backend/requirements.txt
Django==1.7.1
Markdown==2.5.2
PyYAML==3.11
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==0.6
django-extensions==1.4.9
django-filter==0.9.1
django-fsm==2.2.0
django-geoposition==0.2.2
django-rest-swagger==0.2.7
django-toolbelt==0.0.1
djangorestframework==3.0.2
graphviz==0.4.2
gunicorn==19.1.1
psycopg2==2.5.4
py==1.4.26
pytest==2.6.4
pytest-django==2.7.0
pytest-pythonpath==0.5
six==1.9.0
static3==0.5.1
wsgiref==0.1.2
